Upgrade types, and version server.

+++ rtz/src/wasm.rs
@@ -13,7 +13,7 @@ static ALLOC: wee_alloc::WeeAlloc<'_> = wee_alloc::WeeAlloc::INIT;
 /// Get the time zone for the given `(lng,lat)`.
 #[cfg(feature = "tz-ned")]
 #[wasm_bindgen(js_name = getTimezoneNed)]
-pub fn get_timezone_ned(lng: f64, lat: f64) -> JsValue {
+pub fn get_timezone_ned(lng: f32, lat: f32) -> JsValue {
     match crate::get_timezone_ned(lng, lat) {
         Some(tz) => JsValue::from_str(&serde_json::to_string(&tz).unwrap()),
         None => JsValue::NULL,
